Compensated for rotation movement
Even though the sections should be being rotated around the center of the section and thus leaving the center unmoved we found they were actually being moved by quite a substantial amount (~ 10 microns). This looks tricky to fix so instead I measure the depth before rotation and after and then move the section such that its center is in the original position.


Former-commit-id: 32a672159211419d563fe472d31b680fdb3867da